Add the necessary information and plugins for deploying to Sonatype's repositories.
Note that the javadoc plugin is included in the profiles section. This is due to an issue where Java 8 does stricter checking and fails builds. This is mitigated by having a separate profile for Java 8 that removes those strict checks. Source: http://jdpgrailsdev.github.io/blog/2014/04/03/maven_java8_javadoc.html.