Add support for the Lookup API
Every method is added twice, one with the required object, and one with only required parameters.

Version is bumped to 1.2.0 for new functionality.

Version is also bumped for the examples. Bumped to 1.2.0 (instead of 1.1.0) as it should've been bumped when the Verify API was added in a previous pull request.